Chore: Update functionality to unset chapterable selection flag
The functionality to unset the `force_chapterable_selection` flag will
be moved from the dashboard controllers to the chapterable account
assignments controller.

The onboarding requirements on the dashboard controller were
interfering with the unsetting functionality.

Refs: #5694
