+++ app/controllers/mentor/dashboards_controller.rb
@@ -1,7 +1,6 @@
 module Mentor
   class DashboardsController < MentorController
     include LocationStorageController
-    include UnsetForceChapterableSelection
 
     layout "mentor_rebrand"
