+++ app/controllers/student/dashboards_controller.rb
@@ -3,7 +3,6 @@ module Student
     include RequireParentalConsentSigned
     include RequireLocationIsSet
     include LocationStorageController
-    include UnsetForceChapterableSelection
 
     def show
       @regional_events = available_regional_events
