+++ app/views/chapterable_account_assignments/new.en.html.erb
@@ -34,7 +34,7 @@
 
             <div>
               <%= link_to "Confirm",
-                send("unset_force_chapterable_selection_#{current_account.scope_name}_dashboard_path"),
+                unset_force_chapterable_selection_path,
                 method: :post,
                 class: "tw-green-btn"
               %>
