+++ spec/controllers/student/dashboards_controller_spec.rb
@@ -53,20 +53,4 @@ RSpec.describe Student::DashboardsController do
       }
     end
   end
-
-  describe "POST #unset_force_chapterable_selection" do
-    let(:student_profile) { FactoryBot.create(:student) }
-
-    before do
-      student_profile.account.update(force_chapterable_selection: true)
-
-      sign_in(student_profile)
-    end
-
-    it "unsets the `force_chapterable_selection` flag" do
-      post :unset_force_chapterable_selection
-
-      expect(student_profile.account.reload.force_chapterable_selection).to eq(false)
-    end
-  end
 end
